requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
httpx==0.27.2
pydantic==2.11.9
pydantic-settings==2.11.0
PyJWT==2.10.1
supabase==2.4.5
